chore: Refactor to allow easier unit testing
WIP - Proof of Concept
