+++ index.ts
@@ -1,89 +1,20 @@
-import {writable as internal, type Writable} from 'svelte/store'
+import { Writable } from "svelte/store";
+import { type Options, persistedFactory, prodGetStorage, prodIsBrowser } from "./persisted-factory.js";
 
-declare type Updater<T> = (value: T) => T;
-declare type StoreDict<T> = { [key: string]: Writable<T> }
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-interface Stores {
-  local: StoreDict<any>,
-  session: StoreDict<any>,
-}
-
-const stores : Stores = {
-  local: {},
-  session: {}
-}
-
-export interface Serializer<T> {
-  parse(text: string): T
-  stringify(object: T): string
-}
-
-export type StorageType = 'local' | 'session'
-
-export interface Options<T> {
-  serializer?: Serializer<T>
-  storage?: StorageType
-}
-
-function getStorage(type: StorageType) {
-  return type === 'local' ? localStorage : sessionStorage
-}
+/**
+ * Factory function that creates Svelte writable stores backed up by persistent storage.  See the 
+ * [documentation](https://github.com/joshnuss/svelte-persisted-store) for more details.
+ * @param key {string} The key (unique name) to give to the value inside the persisted storage.
+ * @param initialValue {T} The store's initial value.  It is used on first storage read if the storage doesn't have a 
+ * value to offer for the specified key.
+ * @param options {Options<T>} Optional options to fine tune the store's behavior.
+ * @return {Writable<T>} A new Svelte writable store backed up by persisted storage.
+ * @template T
+ */
+export const persisted = persistedFactory(prodGetStorage, prodIsBrowser);
 
 /** @deprecated `writable()` has been renamed to `persisted()` */
 export function writable<T>(key: string, initialValue: T, options?: Options<T>): Writable<T> {
   console.warn("writable() has been deprecated. Please use persisted() instead.\n\nchange:\n\nimport { writable } from 'svelte-persisted-store'\n\nto:\n\nimport { persisted } from 'svelte-persisted-store'")
   return persisted<T>(key, initialValue, options)
-}
-export function persisted<T>(key: string, initialValue: T, options?: Options<T>): Writable<T> {
-  const serializer = options?.serializer ?? JSON
-  const storageType = options?.storage ?? 'local'
-  const browser = typeof(window) !== 'undefined' && typeof(document) !== 'undefined'
-  const storage = browser ? getStorage(storageType) : null
-
-  function updateStorage(key: string, value: T) {
-    storage?.setItem(key, serializer.stringify(value))
-  }
-
-  if (!stores[storageType][key]) {
-    const store = internal(initialValue, (set) => {
-      const json = storage?.getItem(key)
-
-      if (json) {
-        set(<T>serializer.parse(json))
-      }
-
-      if (browser && storageType == 'local') {
-        const handleStorage = (event: StorageEvent) => {
-          if (event.key === key)
-            set(event.newValue ? serializer.parse(event.newValue) : null)
-        }
-
-        window.addEventListener("storage", handleStorage)
-
-        return () => window.removeEventListener("storage", handleStorage)
-      }
-    })
-
-    const {subscribe, set} = store
-
-    stores[storageType][key] = {
-      set(value: T) {
-        updateStorage(key, value)
-        set(value)
-      },
-      update(callback: Updater<T>) {
-        return store.update((last) => {
-          const value = callback(last)
-
-          updateStorage(key, value)
-
-          return value
-        })
-      },
-      subscribe
-    }
-  }
-
-  return stores[storageType][key]
 }
