+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["index.ts"],
+  "include": ["index.ts", "persisted-factory.ts"],
   "exclude": ["node_modules/*"],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
